little improvements

+++ FX/Client.hpp
@@ -11,8 +11,8 @@ class Client
 		std::string _request;
 		std::string _response;
 		int _socket_mod;
-		int _size_response;
-		int _size_request;
+		unsigned long _size_response;
+		unsigned long _size_request;
 
 		const ServerConfig	_server_config;
 
@@ -26,11 +26,11 @@ class Client
 
 		ServerConfig get_server_config() const;
 
-		int get_size_response() const;
-		void set_size_response (int size_response);
+		unsigned long get_size_response() const;
+		void set_size_response (unsigned long size_response);
 
-		int get_size_request() const;
-		void set_size_request (int size_request);
+		unsigned long get_size_request() const;
+		void set_size_request (unsigned long size_request);
 
 		int get_socket_mod() const;
 		void set_socket_mod (int socket_mod);
